+++ src/formatNumberToFixed.ts
@@ -3,33 +3,21 @@ import { Locale } from "./types/locale";
 import { NumberFormat } from "./types/numberFormat";
 import { NumberFormatFixed } from "./types/numberFormatFixed";
 
-const defaultOptions: Partial<NumberFormat> = { 
-  minPrecision: 2,
-  maxPrecision: 2
-}
-
 export function formatNumberToFixed(value: number, precision?: number, locale?: Locale): string
 export function formatNumberToFixed(value: number, options?: Partial<NumberFormatFixed>, locale?: Locale): string
 export function formatNumberToFixed(value: number, options?: Partial<NumberFormatFixed> | number, locale?: Locale): string {
   const isNumberArg = options !== undefined && typeof options === "number",
-    fixedPrecision = isNumberArg ? options : options?.precision ?? defaultOptions.minPrecision,
-    thousandsSeparator = isNumberArg ? undefined : options?.thousandsSeparator,
-    useBankersRounding = isNumberArg ? undefined : options?.useBankersRounding,
-    negativeZero = isNumberArg ? undefined : options?.negativeZero;
+    fixedPrecision = isNumberArg ? options : options?.precision ?? 2 /* HINT: default */;
 
-  // HINT: micro optimization to don't allocate memory when not needed
-  const formatOptions = (fixedPrecision === defaultOptions.minPrecision 
-    && thousandsSeparator === undefined 
-    && useBankersRounding === undefined
-    && negativeZero === undefined)
-      ? defaultOptions 
-      : { 
-      minPrecision: fixedPrecision,
-      maxPrecision: fixedPrecision,
-      thousandsSeparator: thousandsSeparator,
-      useBankersRounding: useBankersRounding,
-      negativeZero: negativeZero
-    };
-
+  const formatOptions: Partial<NumberFormat> = { 
+    minPrecision: fixedPrecision,
+    maxPrecision: fixedPrecision,
+    thousandsSeparator: isNumberArg ? undefined : options?.thousandsSeparator,
+    useBankersRounding: isNumberArg ? undefined : options?.useBankersRounding,
+    negativeZero: isNumberArg ? undefined : options?.negativeZero,
+    currency: isNumberArg ? undefined : options?.currency,
+    currencyDisplay: isNumberArg ? undefined : options?.currencyDisplay,
+  };
+  
   return formatNumber(value, formatOptions, locale);
 }