+++ src/types/numberFormatMoney.ts
@@ -1,10 +1,8 @@
 import { IntlCurrencyDisplay } from "./intlCurrencyDisplay";
 import { Iso4217Alpha3Code } from "./iso4217";
 
-export interface NumberFormatBase {
-  thousandsSeparator: boolean;
-  useBankersRounding: boolean;
-  negativeZero: boolean;
+export interface NumberFormatMoney {
+  precision: number;
   currency: Iso4217Alpha3Code;
   currencyDisplay: IntlCurrencyDisplay;
 }